Create writer file only if metrics are present

+++ src/main/java/org/opensearch/performanceanalyzer/reader_writer_shared/EventLogFileHandler.java
@@ -42,6 +42,8 @@ import java.nio.file.StandardOpenOption;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.opensearch.performanceanalyzer.collectors.StatExceptionCode;
+import org.opensearch.performanceanalyzer.collectors.StatsCollector;
 import org.opensearch.performanceanalyzer.core.Util;
 import org.opensearch.performanceanalyzer.reader.EventDispatcher;
 
@@ -113,11 +115,18 @@ public class EventLogFileHandler {
     public void renameFromTmpWithPrivilege(long epoch) {
         Path path = Paths.get(metricsLocation, String.valueOf(epoch));
         Path tmpPath = Paths.get(path.toString() + TMP_FILE_EXT);
-        // This is done only when no exception is thrown.
-        try {
-            Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
-        } catch (IOException e) {
-            LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+        File tempFile = new File(tmpPath.toString());
+        // Only if the tmp file is present, we want to create the writer file.
+        // If not, we will publish a metric.
+        if (tempFile.exists()) {
+            // This is done only when no exception is thrown.
+            try {
+                Files.move(tmpPath, path, REPLACE_EXISTING, ATOMIC_MOVE);
+            } catch (IOException e) {
+                LOG.error("Error moving file {} to {}.", tmpPath.toString(), path.toString(), e);
+            }
+        } else {
+            StatsCollector.instance().logException(StatExceptionCode.WRITER_FILE_NOT_CREATED_ERROR);
         }
     }
 
@@ -131,7 +140,7 @@ public class EventLogFileHandler {
         File tempFile = new File(pathToFile.toString());
         if (!tempFile.exists()) {
             long mCurrT = System.currentTimeMillis();
-            LOG.info("Didnt find {} at {}", filename, mCurrT);
+            LOG.debug("Didnt find {} at {}", filename, mCurrT);
             return;
         }
         readInternal(pathToFile, BUFFER_SIZE, processor);
